requirements.txt
openai>=1.0.0
langfuse>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
slowapi>=0.1.9
